# src/main/resources/static/swagger-config.yml
openapi: 3.0.1
info:
  title: Gym Service
  version: v0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /api/v1/trainer:
    get:
      tags:
        - Trainer service
      summary: Get trainer by username
      operationId: getTrainerByUsername
      parameters:
        - name: username
          in: query
          required: true
          example: John.Trainer
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerWithTraineesDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainer with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Trainer service
      summary: Update trainer
      operationId: updateTrainer
      requestBody:
        description: New trainer details
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  description: Trainer username (required)
                  type: string
                  example: John.Trainer
                firstName:
                  description: Trainer firstName (required)
                  type: string
                  example: John
                lastName:
                  description: Trainer lastName (required)
                  type: string
                  example: Trainer
                specialization:
                  description: Trainer specialization (required)
                  type: object
                  properties:
                    trainingTypeName:
                      description: Training type name
                      type: string
                      example: Test
                  required:
                    - trainingTypeName
                isActive:
                  description: Trainer isActive status (required)
                  type: boolean
                  example: true
              required:
                - userName
                - firstName
                - lastName
                - specialization
                - isActive
        required: true
      responses:
        '200':
          description: Trainer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerWithTraineesDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainer with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Trainer service
      summary: Create a new trainer
      operationId: createTrainer
      requestBody:
        description: Trainer details
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  description: Trainer firstName (required)
                  type: string
                  example: John
                lastName:
                  description: Trainer lastName (required)
                  type: string
                  example: Trainer
                specialization:
                  description: Trainer specialization (required)
                  type: object
                  properties:
                    trainingTypeName:
                      description: Training type name
                      type: string
                      example: Test
                  required:
                    - trainingTypeName
              required:
                - firstName
                - lastName
                - specialization
        required: true
      responses:
        '201':
          description: Trainer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCredentialsDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Trainer service
      summary: Change trainer status
      operationId: changeTrainerStatus
      parameters:
        - name: username
          in: query
          description: Trainer username
          required: true
          example: John.Trainer
          schema:
            type: string
        - name: isActive
          in: query
          description: Change isActive status for trainer
          required: true
          example: true
          schema:
            type: boolean
      responses:
        '200':
          description: Trainer status changed successfully
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainer with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/trainer/trainings:
    get:
      tags:
        - Trainer service
      summary: Get trainer trainings by criteria
      description: This can only be executed after authentication
      operationId: getTrainerTrainingsByCriteria
      parameters:
        - name: username
          in: query
          required: true
          example: John.Trainer
          description: The trainer username to be provided
          schema:
            type: string
        - name: fromDate
          in: query
          required: false
          example: 2024-01-01
          description: The filtering criteria by date, starting from this specific date
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: false
          example: 2024-02-01
          description: The filtering criteria by date, ending with this specific date
          schema:
            type: string
            format: date
        - name: traineeUsername
          in: query
          required: false
          example: Nick.Trainee
          description: The filtering criteria by trainee username
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainer with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    TrainingTypeDTO:
      required:
        - trainingTypeName
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        trainingTypeName:
          type: string
          description: trainingTypeName (required)
    TraineeDTO:
      required:
        - firstName
        - lastName
      type: object
      properties:
        userName:
          type: string
          description: User userName (optional)
        firstName:
          type: string
          description: User firstName (required)
        lastName:
          type: string
          description: User lastName (required)
        isActive:
          type: boolean
          description: User isActive (optional)
        address:
          type: string
        dateOfBirth:
          type: string
          format: date
    TrainerWithTraineesDTO:
      required:
        - firstName
        - lastName
        - specialization
      type: object
      properties:
        userName:
          type: string
          description: User userName (optional)
        firstName:
          type: string
          description: User firstName (required)
        lastName:
          type: string
          description: User lastName (required)
        isActive:
          type: boolean
          description: User isActive (optional)
        specialization:
          $ref: '#/components/schemas/TrainingTypeDTO'
        trainees:
          type: array
          items:
            $ref: '#/components/schemas/TraineeDTO'
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: message
        errorType:
          type: string
          description: Error type
          example: Type based on error
          enum:
            - VALIDATION_ERROR
            - PROCESSING_ERROR
            - DATABASE_ERROR
            - FATAL_ERROR
            - AUTHORIZATION_ERROR
        timeStamp:
          type: string
          format: date-time
...
    TrainingDTO:
      required:
        - duration
        - trainingDate
        - trainingName
        - trainingTypeName
      type: object
      properties:
        traineeUserName:
          type: string
        trainerUserName:
          type: string
        trainingName:
          type: string
        trainingDate:
          type: string
          format: date
        duration:
          type: integer
          format: int32
        trainingTypeName:
          type: string
    UserCredentialsDTO:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string